aws-cw-orphan-alarms/requirements.txt
boto3==1.34.64
PyYAML==6.0.1
XlsxWriter==3.2.0